Make fingerprint an integration
